/syllabuses/assessments/{syllabus_id} POST started (bad code, will be reworked)
