+++ src/main/java/com/ua/itclusterjava2024/controller/SyllabusesController.java
@@ -1,13 +1,10 @@
 package com.ua.itclusterjava2024.controller;
 
-import com.ua.itclusterjava2024.dto.AssessmentDTO;
 import com.ua.itclusterjava2024.dto.DisciplinesDTO;
 import com.ua.itclusterjava2024.dto.SyllabusesDTO;
-import com.ua.itclusterjava2024.entity.Assessment;
 import com.ua.itclusterjava2024.entity.Disciplines;
 import com.ua.itclusterjava2024.entity.Syllabuses;
 import com.ua.itclusterjava2024.exceptions.NotFoundException;
-import com.ua.itclusterjava2024.service.interfaces.AssessmentService;
 import com.ua.itclusterjava2024.service.interfaces.DisciplinesService;
 import com.ua.itclusterjava2024.service.interfaces.SyllabusesService;
 import com.ua.itclusterjava2024.wrappers.PageWrapper;
@@ -15,12 +12,9 @@ import com.ua.itclusterjava2024.wrappers.Patcher;
 import jakarta.persistence.EntityManager;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 @RestController
 @RequestMapping("/syllabuses")
@@ -28,17 +22,15 @@ public class SyllabusesController {
 
     private final SyllabusesService syllabusesService;
     private final DisciplinesService disciplinesService;
-    private final AssessmentService assessmentService;
     private final ModelMapper modelMapper;
     private final EntityManager entityManager;
     private final Patcher<Syllabuses> patcher;
 
     @Autowired
-    public SyllabusesController(SyllabusesService syllabusesService, AssessmentService assessmentService, DisciplinesService disciplinesService, ModelMapper modelMapper, EntityManager entityManager, Patcher<Syllabuses> patcher) {
+    public SyllabusesController(SyllabusesService syllabusesService, DisciplinesService disciplinesService, ModelMapper modelMapper, EntityManager entityManager, Patcher<Syllabuses> patcher) {
         this.syllabusesService = syllabusesService;
         this.disciplinesService = disciplinesService;
         this.modelMapper = modelMapper;
-        this.assessmentService = assessmentService;
         this.entityManager = entityManager;
         this.patcher = patcher;
     }
@@ -87,19 +79,6 @@ public class SyllabusesController {
     public PageWrapper<SyllabusesDTO> delete(@PathVariable Long id) {
         syllabusesService.delete(id);
         return findAll();
-    }
-
-    @GetMapping("/assessments/{syllabus_id}")
-    public ResponseEntity<?> getAssessmentsBySyllabusId(@PathVariable("syllabus_id") Long syllabusId) {
-        List<AssessmentDTO> assessmentsDTO = assessmentService.getAllAssessmentsBySyllabus(syllabusId)
-                .stream().map((element) -> modelMapper.map(element, AssessmentDTO.class))
-                .toList();
-        Map<String, Object> content = new HashMap<>();
-        content.put("assessments", assessmentsDTO);
-        content.put("syllabus_id", syllabusId);
-        Map<String, Object> response = new HashMap<>();
-        response.put("content", content);
-        return ResponseEntity.ok(response);
     }
 
     private Syllabuses convertToEntity(SyllabusesDTO syllabusesDTO) {
