+++ src/main/java/com/ua/itclusterjava2024/repository/AssessmentRepository.java
@@ -2,8 +2,7 @@ package com.ua.itclusterjava2024.repository;
 
 import com.ua.itclusterjava2024.entity.Assessment;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
+
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
