+++ src/main/java/com/ua/itclusterjava2024/service/implementation/AssessmentServiceImpl.java
@@ -2,7 +2,9 @@ package com.ua.itclusterjava2024.service.implementation;
 
 import com.ua.itclusterjava2024.entity.Assessment;
 import com.ua.itclusterjava2024.repository.AssessmentRepository;
+import com.ua.itclusterjava2024.repository.UserRepository;
 import com.ua.itclusterjava2024.service.interfaces.AssessmentService;
+import com.ua.itclusterjava2024.service.interfaces.SyllabusesService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -13,9 +15,11 @@ import java.util.Optional;
 @Service
 public class AssessmentServiceImpl implements AssessmentService {
     private final AssessmentRepository assessmentRepository;
+    private final SyllabusesService syllabusesService;
 
-    public AssessmentServiceImpl(AssessmentRepository assessmentRepository) {
+    public AssessmentServiceImpl(AssessmentRepository assessmentRepository, SyllabusesService syllabusesService) {
         this.assessmentRepository = assessmentRepository;
+        this.syllabusesService = syllabusesService;
     }
 
     @Override
@@ -49,8 +53,11 @@ public class AssessmentServiceImpl implements AssessmentService {
         return assessmentRepository.findAll(pageable);
     }
 
-
     public List<Assessment> getAllAssessmentsBySyllabus(long syllabusId) {
         return assessmentRepository.findAllBySyllabusId(syllabusId);
     }
+
+//    public List<Assessment> createAllAssessmentsWithSyllabus(List<Assessment> assessments, long syllabusId) {
+//
+//    }
 }
